Add tests for MainActivity and BoredActivity
