+++ app/src/androidTest/java/com/github/sdpcoachme/MainActivityTest.java
@@ -9,6 +9,10 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
+import android.content.Intent;
+
+import androidx.test.core.app.ActivityScenario;
+import androidx.test.core.app.ApplicationProvider;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
@@ -28,7 +32,7 @@ public class MainActivityTest {
     public ActivityScenarioRule<MainActivity> testRule = new ActivityScenarioRule<>(MainActivity.class);
 
     @Test
-    public void testButton(){
+    public void okButton(){
         String str = "Enter your name";
         Intents.init();
 
@@ -41,4 +45,17 @@ public class MainActivityTest {
 
         Intents.release();
     }
+
+    @Test
+    public void APIPageButton(){
+        Intents.init();
+
+        onView(withId(R.id.boredAPI)).check(matches(isDisplayed()));
+        onView(withId(R.id.boredAPI)).perform(click());
+
+        Intents.intended(hasComponent(BoredActivity.class.getName()));
+
+        Intents.release();
+    }
+
 }
